tech-job-board/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
langchain==0.0.352
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.7.2
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
httpx==0.26.0
beautifulsoup4==4.12.3
requests==2.31.0
scikit-learn==1.4.0
numpy==1.26.3
sentence-transformers==2.3.1
